+++ src/main/java/roomescape/auth/config/JwtConfig.java
@@ -1,0 +1,19 @@
+package roomescape.auth.config;
+
+import io.jsonwebtoken.security.Keys;
+import javax.crypto.SecretKey;
+import lombok.RequiredArgsConstructor;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+@RequiredArgsConstructor
+public class JwtConfig {
+
+    private final JwtProperties jwtProperties;
+
+    @Bean
+    public SecretKey jwtKey() {
+        return Keys.hmacShaKeyFor(jwtProperties.getSecretKey().getBytes());
+    }
+}
