+++ src/main/java/roomescape/auth/config/JwtProperties.java
@@ -1,0 +1,15 @@
+package roomescape.auth.config;
+
+import lombok.Getter;
+import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.stereotype.Component;
+
+@Component
+@ConfigurationProperties(prefix = "jwt")
+@Getter
+public class JwtProperties {
+
+    private String secretKey;
+    private int expiration;
+    private String issuer;
+}
