+++ src/main/java/roomescape/auth/web/controller/AuthController.java
@@ -1,4 +1,4 @@
-package roomescape.auth.controller;
+package roomescape.auth.web.controller;
 
 import static roomescape.auth.web.constant.AuthConstant.AUTH_COOKIE_KEY;
 
@@ -12,7 +12,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.auth.controller.dto.LoginRequest;
+import roomescape.auth.web.controller.dto.LoginRequest;
 import roomescape.auth.dto.AuthenticatedMember;
 import roomescape.auth.application.AuthService;
 import roomescape.auth.web.support.CookieAuthorizationExtractor;
